% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_function.R
\name{median_function2024}
\alias{median_function2024}
\title{This function allows you to calculate the median from a numeric vector.}
\usage{
median_function2024(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to calculate the median from a numeric vector.
}
\examples{
median_function2024(seq(1:10))
}
\keyword{median}
